move frame, series back
