--- pandas/core/arrays/sparse/api.py
@@ -1,6 +1,0 @@
-# pylint: disable=W0611
-# flake8: noqa
-from pandas.core.arrays.sparse.array import SparseArray
-from pandas.core.arrays.sparse.series import SparseSeries
-from pandas.core.arrays.sparse.frame import SparseDataFrame
-from pandas.core.arrays.sparse.dtype import SparseDtype
