+++ pandas/core/reshape/reshape.py
@@ -18,8 +18,8 @@ from pandas.core.dtypes.missing import notna
 from pandas.core.series import Series
 from pandas.core.frame import DataFrame
 
-from pandas.core.arrays.sparse.api import SparseDataFrame, SparseSeries
-from pandas.core.arrays.sparse.array import SparseArray
+from pandas.core.sparse.api import SparseDataFrame, SparseSeries
+from pandas.core.arrays import SparseArray
 from pandas._libs.sparse import IntIndex
 
 from pandas.core.arrays import Categorical
