+++ pandas/core/sparse/api.py
@@ -1,0 +1,5 @@
+# pylint: disable=W0611
+# flake8: noqa
+from pandas.core.arrays.sparse import SparseArray, SparseDtype
+from pandas.core.sparse.series import SparseSeries
+from pandas.core.sparse.frame import SparseDataFrame
