+++ pandas/core/sparse/frame.py
@@ -22,8 +22,8 @@ import pandas.core.algorithms as algos
 from pandas.core.internals import (BlockManager,
                                    create_block_manager_from_arrays)
 import pandas.core.generic as generic
-from pandas.core.arrays.sparse.series import SparseSeries, SparseArray
-from pandas.core.arrays.sparse.dtype import SparseDtype
+from pandas.core.arrays.sparse import SparseArray, SparseDtype
+from pandas.core.sparse.series import SparseSeries
 from pandas._libs.sparse import BlockIndex, get_blocks
 from pandas.util._decorators import Appender
 import pandas.core.ops as ops
