+++ pandas/io/packers.py
@@ -69,8 +69,8 @@ from pandas.core import internals
 from pandas.core.arrays import IntervalArray
 from pandas.core.generic import NDFrame
 from pandas.core.internals import BlockManager, make_block, _safe_reshape
-from pandas.core.arrays.sparse.api import SparseSeries, SparseDataFrame
-from pandas.core.arrays.sparse.array import BlockIndex, IntIndex
+from pandas.core.sparse.api import SparseSeries, SparseDataFrame
+from pandas.core.arrays.sparse import BlockIndex, IntIndex
 from pandas.io.common import get_filepath_or_buffer, _stringify_path
 from pandas.io.msgpack import Unpacker as _Unpacker, Packer as _Packer, ExtType
