+++ pandas/tests/arrays/sparse/test_arithmetics.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 import pandas as pd
 import pandas.util.testing as tm
-from pandas.core.arrays.sparse.api import SparseDtype
+from pandas.core.sparse.api import SparseDtype
 
 
 class TestSparseArrayArithmetics(object):
