+++ pandas/tests/series/test_subclass.py
@@ -2,7 +2,7 @@
 # pylint: disable-msg=E1101,W0612
 import numpy as np
 import pandas as pd
-from pandas.core.arrays.sparse.dtype import SparseDtype
+from pandas import SparseDtype
 import pandas.util.testing as tm
