+++ pandas/tests/sparse/frame/test_analytics.py
