+++ pandas/tests/sparse/frame/test_indexing.py
