+++ pandas/tests/sparse/series/test_indexing.py
