+++ pandas/tests/sparse/test_format.py
