+++ pandas/tests/sparse/test_groupby.py
