+++ pandas/tests/sparse/test_indexing.py
@@ -4,7 +4,7 @@ import pytest
 import numpy as np
 import pandas as pd
 import pandas.util.testing as tm
-from pandas.core.arrays.sparse.api import SparseDtype
+from pandas.core.sparse.api import SparseDtype
 
 
 class TestSparseSeriesIndexing(object):
