+++ pandas/tests/sparse/test_pivot.py
