+++ pandas/tests/sparse/test_reshape.py
